+++ ts/src/Group.ts
@@ -1,3 +1,4 @@
 export * from "./GroupMeta";
 export * from "./GroupChannel";
 export * from "./GroupManager";
+export * from "./SenderKeyRepair";